fix: Spring Security + Jwt 적용 후, 로그인된 유저에 한해서, 서비스가 작동되도록 코드 리팩토링

+++ src/main/java/self/study/controller/BoardController.java
@@ -3,7 +3,9 @@ package self.study.controller;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
+import self.study.config.auth.PrincipalDetails;
 import self.study.dto.BoardDto;
 import self.study.entity.User;
 import self.study.repository.UserRepository;
@@ -27,7 +29,6 @@ public class BoardController {
     }
 
 
-
     // 개별 게시글 조회
     @ApiOperation(value = "개별 게시글 보기", notes = "개별 게시글 조회한다.")
     @ResponseStatus(HttpStatus.OK)
@@ -37,26 +38,22 @@ public class BoardController {
     }
 
 
-
     // 게시글 작성
     @ApiOperation(value = "게시글 작성", notes = "게시글을 작성한다.")
     @ResponseStatus(HttpStatus.CREATED)
     @PostMapping("/boards/write")
-    public Response write(@RequestBody BoardDto boardDto) {
-        // 원래 로그인을 하면, User 정보는 세션을 통해서 구하고 주면 되지만,
-        // 지금은 핵심 개념을 알기 위해서, JWT 로그인은 생략하고, 임의로 findById 로 유저 정보를 넣어줬습니다.
-
-        User user = userRepository.findById(1).get();
+    public Response write(@RequestBody BoardDto boardDto, Authentication authentication) {
+        PrincipalDetails principalDetails = (PrincipalDetails) authentication.getPrincipal();
+        User user = principalDetails.getUser();
         return new Response("성공", "글 작성 성공", boardService.write(boardDto, user));
     }
-
 
 
     // 게시글 수정
     @ApiOperation(value = "게시글 수정", notes = "게시글을 수정한다.")
     @ResponseStatus(HttpStatus.OK)
     @PutMapping("/boards/update/{id}")
-    public Response edit(@RequestBody BoardDto boardDto, @PathVariable("id") Integer id) {
+    public Response edit(@RequestBody BoardDto boardDto, @PathVariable("id") Integer id, Authentication authentication) {
         // 원래 로그인을 하면, User 정보는 세션을 통해서 구하고 주면 되지만,
         // 지금은 핵심 개념을 알기 위해서, JWT 로그인은 생략하고, 임의로 findById 로 유저 정보를 넣어줬습니다.
 
@@ -67,22 +64,34 @@ public class BoardController {
         // 맞으면 아래 로직 수행, 틀리면 다른 로직(ResponseFail 등 커스텀으로 만들어서) 수행
         // 이건 if문으로 처리할 수 있습니다. * 이 방법 말고 service 내부에서 확인해도 상관 없음
 
-        User user = userRepository.findById(1).get();
-        return new Response("성공", "글 수정 성공", boardService.update(id, boardDto));
+        PrincipalDetails principalDetails = (PrincipalDetails) authentication.getPrincipal();
+        User user = principalDetails.getUser();
+        if (user.getName().equals(boardService.getBoard(id).getWriter())) {
+            // 로그인된 유저의 글이 맞다면
+            return new Response("성공", "글 수정 성공", boardService.update(id, boardDto));
+        } else {
+            return new Response("실패", "본인 게시물만 수정할 수 있습니다.", null);
+
+        }
     }
-
-
 
 
     // 게시글 삭제
     @ApiOperation(value = "게시글 삭제", notes = "게시글을 삭제한다.")
     @ResponseStatus(HttpStatus.OK)
     @DeleteMapping("/boards/delete/{id}")
-    public Response delete(@PathVariable("id") Integer id) {
+    public Response delete(@PathVariable("id") Integer id, Authentication authentication) {
         // 이것도 마찬가지로, JWT(로그인 관련) 공부를 하고나서 현재 이 요청을 보낸 로그인된 유저의 정보가
         // 게시글의 주인인지 확인하고, 맞으면 삭제 수행 후 리턴해주고, 틀리면 에러 리턴해주면 됩니다.
 
-        boardService.delete(id); // 이 메소드는 반환값이 없으므로 따로 삭제 수행해주고, 리턴에는 null을 넣어줌
-        return new Response("성공", "글 삭제 성공", null);
+        PrincipalDetails principalDetails = (PrincipalDetails) authentication.getPrincipal();
+        User user = principalDetails.getUser();
+        if (user.getName().equals(boardService.getBoard(id).getWriter())) {
+            // 로그인된 유저가 글 작성자와 같다면
+            boardService.delete(id); // 이 메소드는 반환값이 없으므로 따로 삭제 수행해주고, 리턴에는 null을 넣어줌
+            return new Response("성공", "글 삭제 성공", null);
+        } else {
+            return new Response("실패", "본인 게시물만 삭제할 수 있습니다.", null);
+        }
     }
 }
